\name{egg_price_index}
\docType{data}
\alias{egg_price_index}
\title{Price Index of Eggs in Urban Ares of India}
\description{
This data set contains observations of Consumer Price Index (CPI) of eggs for urban areas in India. It consists of 86 monthly observations, ranging from January, 2013 till February, 2020. Some observations after this point of time were also available, but were discarded due to their questionable reliability and missingness pertaining to the onset of COVID-19 pandemic from March, 2020.}
\usage{egg_price_index}
\format{A data frame containing 86 observations. Observations are contained in the \code{"egg_price_index"} column, while the first two columns pertain to year and month of observation, repectively.}
\keyword{datasets}
